refactor: extract htmlUtils from tagUtils to break DraftMatecatUtils bilateral cycle
Move getIdAttributeRegEx, unescapeHTMLinTags, and unescapeHTMLRecursive into a new htmlUtils.js file to eliminate the tagUtils → textUtils → tagUtils cycle. tagUtils re-exports for backward compatibility.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
